+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/datasources/samples/SamplesDbConnector.groovy
@@ -9,6 +9,7 @@ import life.qbic.portal.sampletracking.datasources.database.ConnectionProvider
 
 import java.sql.Connection
 import java.sql.PreparedStatement
+import java.sql.Statement
 import java.sql.ResultSet
 
 /**