+++ CHANGELOG.rst
@@ -14,7 +14,11 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 * Count of samples that finished library prep is shown (`#89 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/89>`_)
 
+* Enable user to subscribe to individual projects (`#84 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/84>`_)
+
 **Fixed**
+
+* Remove possibility of duplicate subscription of user per project (`#95 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/95>`_)
 
 **Dependencies**